modularized ipchandler
